requirements.txt
python-dotenv~=0.21.0
python-telegram-bot==13.15
vk_api==11.9.*
